+++ include/asbind20/debugging.hpp
@@ -15,6 +15,7 @@
 
 #include "debugging/extract_string.hpp"
 #include "debugging/gc_statistics.hpp"
+#include "debugging/source_location.hpp"
 
 // IWYU pragma: end_exports
 
@@ -48,58 +49,6 @@ inline const char* get_function_section_name(
     return func->GetScriptSectionName();
 #endif
 }
-
-struct script_source_location
-{
-    using value_type = int;
-
-    cstring_ref section_name{};
-    cstring_ref function_name{};
-    int line = 0;
-    int column = 0;
-
-    constexpr script_source_location() noexcept = default;
-    constexpr script_source_location(const script_source_location&) noexcept = default;
-
-    script_source_location& operator=(const script_source_location&) noexcept = default;
-
-#if ANGELSCRIPT_VERSION >= 23800
-
-    static script_source_location from_function(const_function_pointer func)
-    {
-        script_source_location result{};
-        if(!func) [[unlikely]]
-            return result;
-
-        const char* section = nullptr;
-        func->GetDeclaredAt(&section, &result.line, &result.column);
-        result.section_name = section;
-        result.function_name = func->GetName();
-
-        return result;
-    }
-
-#endif
-
-    static script_source_location from_context(
-        context_pointer ctx,
-        AS_NAMESPACE_QUALIFIER asUINT stack_level = 0
-    )
-    {
-        script_source_location result{};
-        if(!ctx) [[unlikely]]
-            return result;
-
-        const char* section = nullptr;
-        result.line = ctx->GetLineNumber(stack_level, &result.column, &section);
-        result.section_name = section;
-        auto* func = ctx->GetFunction(stack_level);
-        if(func)
-            result.function_name = func->GetName();
-
-        return result;
-    }
-};
 } // namespace asbind20::debugging
 
 #endif
